Make validate_license a method
